requirements.txt
pyyaml>=6.0
pandas>=2.0
matplotlib>=3.7
seaborn>=0.12
